+++ Runtime/GameEngine/Libraries/Match/MatchTcpClients/Synchronizer/ClientSynchronizer.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
@@ -19,7 +18,7 @@ namespace MatchTcpClients.Synchronizer
 
 		private readonly ClientSynchronizerConfig _config;
 		private          string                   _sessionToken;
-		private          DateTime?                _lastReceivedPingDataTime = null;
+		private          DateTime?                _lastReceivedPingDataTime;
 		private          NtpData                  _lastReceivedUnreliableNtpData;
 		private          bool                     _waitingForFirstUnreliablePing = true;
 
@@ -58,7 +57,7 @@ namespace MatchTcpClients.Synchronizer
 					stopwatch.Reset();
 
 					if (timeToWait > TimeSpan.Zero)
-						await Task.Delay(timeToWait, ct).ContinueWith(_ => { }, CancellationToken.None);
+						await Task.Delay(timeToWait, ct).CatchOperationCanceledException();
 				}
 			}
 		}
@@ -74,7 +73,7 @@ namespace MatchTcpClients.Synchronizer
 			SendSynchronizeRequest();
 
 			var pingCompletionTask = pingCompletionSource.Task;
-			var timeoutTask = Task.Delay(_config.TimeoutTime, ct).ContinueWith(_ => { }, CancellationToken.None);
+			var timeoutTask = Task.Delay(_config.TimeoutTime, ct).CatchOperationCanceledException();
 
 			var firstFinishedTask = await Task.WhenAny(pingCompletionTask, timeoutTask);
 			_pingResponseCallback = null;