+++ Runtime/GameEngine/Libraries/Match/MatchTcpClients/WebGameServerClient.cs
@@ -3,7 +3,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using MatchTcpLibrary;
 using MatchTcpLibrary.TransportLayer.WebRtc;
-using UnityEngine;
 using WebRtcWrapper;
 
 namespace MatchTcpClients
@@ -103,13 +102,15 @@ namespace MatchTcpClients
 				signalingClient.ReceivedResponse += OnAnswerReceived;
 				signalingClient.PostOfferAsync(offer, (int)Math.Ceiling(Config.OfferTimeout.TotalSeconds), linkedCts.Token);
 
-				await Task.Delay(Config.OfferRetryDelay, linkedCts.Token).ContinueWith(_ => { }, CancellationToken.None);
+				await Task.Delay(Config.OfferTimeout, linkedCts.Token).CatchOperationCanceledException();
 				signalingClient.ReceivedResponse -= OnAnswerReceived;
 				cts.Cancel();
 
 				if (response?.IsError == false)
 					break;
 				Logger.Error("[Elympics] WebRTC answer error: {0}", response?.Text);
+
+				await Task.Delay(Config.OfferRetryDelay, linkedCts.Token).CatchOperationCanceledException();
 			}
 
 			return response;
@@ -131,7 +132,7 @@ namespace MatchTcpClients
 
 			_webRtcClient.OfferCreated += OnOfferCreated;
 			_webRtcClient.CreateOffer();
-			await Task.Delay(Config.OfferTimeout, cts.Token).ContinueWith(_ => { }, CancellationToken.None);
+			await Task.Delay(Config.OfferTimeout, cts.Token).CatchOperationCanceledException();
 			_webRtcClient.OfferCreated -= OnOfferCreated;
 			return (offer, offerSet);
 		}